+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
+## 16.07.2022
+### Added
+Support access to non-localhost environments
+Create method that allows insering multiple CSV files
+
+### Changed
+Fix create Pagelinks MATCH to only match the right pages 
+
 ## 26.06.2022
 ### Added
 - support for gds in docker for neo4j